+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * Alt+Return to emit "ESC \r".
 * Trackpad sloooow scrolling to eventually scroll a line.
 * Memory leak in terminal reset.
+* Translation of cursor coordinates on resize
 
 
 ### Security